update entityset tests
